Extract the table filtering code to a common file
